requirements.txt
deta==1.1.0
pillow==9.5.0
flask==2.2.3
fastapi==0.95.1
a2wsgi==1.7.0
